refactor: rename auth-error.ts.js to exceptions.ts and update imports accordingly bb-6
