+++ apps/backend/src/modules/auth/libs/exceptions/exceptions.ts
